fix(rs-sdk): dumping quorum keys test vectors does not work
